Move listTrainingEntries to fineTunes router
