+++ app/src/server/api/routers/datasetEntries.router.ts
@@ -121,64 +121,6 @@ export const datasetEntriesRouter = createTRPCRouter({
         totalTestingCount,
       };
     }),
-  listTrainingEntries: protectedProcedure
-    .input(z.object({ fineTuneId: z.string(), page: z.number(), pageSize: z.number() }))
-    .query(async ({ input, ctx }) => {
-      const { fineTuneId, page, pageSize } = input;
-
-      const fineTune = await prisma.fineTune.findUnique({
-        where: {
-          id: fineTuneId,
-        },
-      });
-
-      if (!fineTune) throw new TRPCError({ message: "Fine tune not found", code: "NOT_FOUND" });
-      await requireCanViewProject(fineTune.projectId, ctx);
-
-      const [entries, count] = await prisma.$transaction([
-        prisma.fineTuneTrainingEntry.findMany({
-          where: {
-            fineTuneId: fineTuneId,
-          },
-          include: {
-            datasetEntry: {
-              select: {
-                messages: true,
-                function_call: true,
-                functions: true,
-                tool_choice: true,
-                tools: true,
-                output: true,
-                inputTokens: true,
-                outputTokens: true,
-              },
-            },
-          },
-          orderBy: {
-            datasetEntry: {
-              sortKey: "desc",
-            },
-          },
-          skip: (page - 1) * pageSize,
-          take: pageSize,
-        }),
-        prisma.fineTuneTrainingEntry.count({
-          where: {
-            fineTuneId: fineTuneId,
-          },
-        }),
-      ]);
-
-      const typedEntries = entries.map((entry) => ({
-        ...entry,
-        datasetEntry: typedDatasetEntry(entry.datasetEntry),
-      }));
-
-      return {
-        entries: typedEntries,
-        count,
-      };
-    }),
   get: protectedProcedure.input(z.object({ id: z.string() })).query(async ({ input, ctx }) => {
     const entry = await prisma.datasetEntry.findUniqueOrThrow({
       where: { id: input.id },
